% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3barplot.R
\name{c3barplot-shiny}
\alias{c3barplot-shiny}
\alias{c3barplotOutput}
\alias{renderC3barplot}
\title{Shiny bindings for c3barplot}
\usage{
c3barplotOutput(outputId, width = "100\%", height = "400px")

renderC3barplot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a c3barplot}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using c3barplot within Shiny
applications and interactive Rmd documents.
}
