% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot-shiny}
\alias{c3plot-shiny}
\alias{c3plotOutput}
\alias{renderC3plot}
\title{Shiny bindings for c3plot}
\usage{
c3plotOutput(outputId, width = "100\%", height = "400px")

renderC3plot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a c3Scatter}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using c3plot within Shiny
applications and interactive Rmd documents.
}
