% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot.default}
\alias{c3plot.default}
\title{The Default C3 Scatterplot Function}
\usage{
\method{c3plot}{default}(
  x,
  y,
  type = "p",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zoom = TRUE,
  col.group = NULL,
  col = NULL,
  ...
)
}
\arguments{
\item{x}{the x coordinates for the plot.}

\item{y}{the y coordinates for the plot.}

\item{type}{1-character string giving the type of plot desired. The following values are possible: "p" for points, "l" for lines, and "b" for both points and lines.}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{zoom}{logical; should the zooming feature (controlled my mouse wheel event) be enabled for the plot?}

\item{col.group}{optionally, a factor the same length as \code{x} by which to group and color points.}

\item{col}{The colors for the lines and points. If \code{col.group} is specified, this can be a vector of colors to use for each group in the data. If \code{NULL}, the C3 default colors are used.}

\item{...}{arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}: \code{width}, \code{height}, and \code{elementId}. These arguments default to NULL.}
}
\description{
Draw an interactive scatterplot or line plot using 'C3.js'.
}
\examples{
data(mtcars)
c3plot(mtcars$hp, mtcars$qsec)
c3plot(mtcars$disp, mtcars$hp, main = "Displacement vs. HP in mtcars")
}
