% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot.density}
\alias{c3plot.density}
\title{c3plot Method for  Kernel Density Estimation}
\usage{
\method{c3plot}{density}(x, main = NULL, xlab = NULL, ylab = "Density", type = "l", ...)
}
\arguments{
\item{x}{a "density" object}

\item{main}{a main title for the plot.}

\item{xlab}{label for the x axis}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{type}{1-character string giving the type of plot desired. The following values are possible: "p" for points, "l" for lines, and "b" for both points and lines.}

\item{...}{arguments passed to other methods}
}
\description{
The \code{c3plot} method for density objects
}
