% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3barplot.R
\name{c3plot.factor}
\alias{c3plot.factor}
\title{Factor Variable Bar Plots}
\usage{
\method{c3plot}{factor}(x, ylab = "Count", ...)
}
\arguments{
\item{x}{a factor.}

\item{ylab}{a label for the y axis.}

\item{...}{arguments passed to \code{\link{c3barplot}}.}
}
\description{
This function implements a bar plot method for \code{factor} arguments to the \code{\link{c3plot}} generic function. Bar heights will be counts for the factor levels.
}
\examples{
mtcars <- mtcars
mtcars$cyl <- as.factor(mtcars$cyl)
c3plot(mtcars$cyl)
}
\seealso{
\code{\link[c3plot:c3barplot]{c3barplot()}}
}
