% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot.function}
\alias{c3plot.function}
\title{Draw Function Plots with 'C3.js'}
\usage{
\method{c3plot}{`function`}(x, from = 0, to = 1, ylab = NULL, ...)
}
\arguments{
\item{x}{The name of a function.}

\item{from}{the lower limit of the range over which the function will be plotted.}

\item{to}{the upper limit of the range over which the function will be plotted.}

\item{ylab}{a label for the y axis, defaults to the name of the function specified in \code{x}.}

\item{...}{arguments passed to other methods}
}
\description{
Draws a curve corresponding to a function over the interval \code{[from, to]} in an interactive plot using 'C3.js'
}
\examples{
c3plot(qnorm)

c3plot(sin, -pi, 2*pi)
}
