% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot.lm}
\alias{c3plot.lm}
\title{C3 Plot Diagnostics for an lm Object}
\usage{
\method{c3plot}{lm}(x, which = 1, ...)
}
\arguments{
\item{x}{an "lm" object}

\item{which}{1 is the only supported value currently. This argument is included for consistency with the base \code{plot.lm()} method.}

\item{...}{arguments passed to other methods.}
}
\description{
a plot of residuals against fitted values
}
\examples{
## Analysis of the life-cycle savings data
## given in Belsley, Kuh and Welsch.
lm.SR <- lm(sr ~ pop15 + pop75 + dpi + ddpi, data = LifeCycleSavings)
c3plot(lm.SR)

}
