% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3Scatter.R
\name{c3plot}
\alias{c3plot}
\title{Generic Plotting with 'C3.js'}
\usage{
c3plot(x, ...)
}
\arguments{
\item{x}{x coordinates for points or any 'R' object with a \code{c3plot} method.}

\item{...}{arguments passed to methods.}
}
\description{
This is a generic function for plotting 'R' objects using the \href{http://c3js.org/}{C3.js} charting library. The syntax is similar to that of the \code{plot()} generic function.
}
\details{
For simple scatter and line plots, \code{\link{c3plot.default}} will be used. However, there are \code{c3plot} methods for various 'R' objects. Use \code{methods(c3plot)} and the documentation for these.


Plots created with this are interactive \code{\link[htmlwidgets:htmlwidgets-package]{htmlwidgets}} that can be used in the RStudio Viewer or embedded into 'R Markdown' documents, 'Shiny' web applications, etc.

For the default method, the argument \code{y} with the y coordinates of points is required. For some methods, such as \code{\link{c3plot.density}} and \code{\link{c3plot.function}}, \code{y} is not required because those methods can compute or extract y coordinates from the \code{x} object
}
\examples{
data(mtcars)
c3plot(mtcars$disp, mtcars$hp, main = "Displacement vs. HP in mtcars")


c3plot(qnorm)
}
