% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjsScatter.R
\name{plotjs.lm}
\alias{plotjs.lm}
\title{plotjs Diagnostics for an lm Object}
\usage{
\method{plotjs}{lm}(x, which = 1, ...)
}
\arguments{
\item{x}{an "lm" object}

\item{which}{1 is the only supported value currently. This argument is included for consistency with the base \code{plot.lm()} method.}

\item{...}{arguments passed to other methods.}
}
\description{
A plot of residuals against fitted values.
}
\examples{
lm.SR <- lm(sr ~ pop15 + pop75 + dpi + ddpi, data = LifeCycleSavings)
plotjs(lm.SR)
}
