% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjsScatter.R
\name{plotjs}
\alias{plotjs}
\title{Generic Plotting with 'Chart.js'}
\usage{
plotjs(x, ...)
}
\arguments{
\item{x}{x coordinates for points or any 'R' object with a \code{plotjs} method.}

\item{...}{arguments passed to methods.}
}
\description{
This is a generic function for plotting 'R' objects using the \href{https://www.chartjs.org/}{Chart.js} charting library. The syntax is similar to that of the \code{plot()} generic function.
}
\details{
For simple scatter and line plots, \code{\link{plotjs.default}} will be used. However, there are \code{plotjs} methods for various 'R' objects. Use \code{methods(plotjs)} and the documentation for these.


Plots created with this are interactive \code{\link[htmlwidgets:htmlwidgets-package]{htmlwidgets}} that can be used in the RStudio Viewer or embedded into 'R Markdown' documents, 'Shiny' web applications, etc.

For the default method, the argument \code{y} with the y coordinates of points is required. For some methods, such as \code{\link{plotjs.density}} and \code{\link{plotjs.function}}, \code{y} is not required because those methods can compute or extract y coordinates from the \code{x} object.
}
\examples{
data(mtcars)
plotjs(mtcars$disp, mtcars$hp, main = "Displacement vs. HP in mtcars")
plotjs(qnorm)
}
